+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [0ver](https://0ver.org).
 
 ## [Unreleased]
 
+## [0.2.12] - 2019-03-09
+
 ### Added
 
 - Now using stretchr/testify for asserting test results
@@ -15,6 +17,8 @@ and this project adheres to [0ver](https://0ver.org).
 ### Changed
 
 - Fix `--gitlab-token` and improve docs/chart
+- Bumped to go 1.14
+- Bumped goreleaser to 0.128.0
 
 ## [0.2.11] - 2019-02-03
 
@@ -188,7 +192,8 @@ and this project adheres to [0ver](https://0ver.org).
 - LICENSE
 - README
 
-[Unreleased]: https://github.com/mvisonneau/gitlab-ci-pipelines-exporter/compare/0.2.11...HEAD
+[Unreleased]: https://github.com/mvisonneau/gitlab-ci-pipelines-exporter/compare/0.2.12...HEAD
+[0.2.12]: https://github.com/mvisonneau/gitlab-ci-pipelines-exporter/tree/0.2.12
 [0.2.11]: https://github.com/mvisonneau/gitlab-ci-pipelines-exporter/tree/0.2.11
 [0.2.10]: https://github.com/mvisonneau/gitlab-ci-pipelines-exporter/tree/0.2.10
 [0.2.9]: https://github.com/mvisonneau/gitlab-ci-pipelines-exporter/tree/0.2.9